pyproject.toml
[tool.poetry]
name = "open-source-code-vulnerabilities-dataset"
version = "0.1.0"
description = ""
authors = ["vodkar <boombarah@gmail.com>"]
packages = [{ include = "open_source_code_vulnerabilities_dataset" }]

[tool.poetry.dependencies]
python = "^3.12"
jupyter-book = "^1.0.2"
matplotlib = "^3.9.2"
numpy = "^2.1.1"
polars = "^1.7.1"
pygithub = "^2.5.0"
ipykernel = "^6.29.5"
ruff = "^0.9.0"
python-gitlab = "^5.3.1"
whatthepatch = "^1.0.7"
gitpython = "^3.1.44"

[tool.poetry.group.finetune.dependencies]
transformers = "^4.48.1"
torch = { version = "^2.5.1", source = "torch" }
xformers = "^0.0.29"
unsloth = "^2025.1.6"
ipywidgets = "^8.1.5"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu124"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = ["repos/"]
